Intermediary work for modularising df-py.
